Show to resolve a Promise
The `flagChain` function returns the result of `getImageName`, which is
a promise. The `then` method lets us implicitly pass the settled (either
resolved or rejected) promise to another function. The `then` method
takes two arguments in the following order:

1. The function to be called if the promise resolves.
2. The function to be called if the promise rejects.

If the first function is called, then it is implicitly passed the
resolved promise value. If the second function is called, then it is
implicitly passed the rejection error.

`.then(logSuccess, logError);`
* Using named functions inside of `then` is a good practice.