Show why using `catch` is a good idea
Notice the suble change of adding `catch`.

* The `catch` method is similar to `then`, but only deals with rejected
cases. With this new pattern, if the promise from `getImageName`
resolves, then we just call `logSuccess` but if it rejects, `logError`
is caught in the `catch`.

This new code also triggers `catch` if `logSuccess` rejects, because
`logSuccess` occurs before the catch.

Promise rejections skip forward to the next `then` with a rejection
callback (or catch since they're equivalent).